pyproject.toml
[tool.poetry]
name = "hazel"
version = "0.1.2"
description = "A reusable toolkit library for desktop and web applications"
authors = ["Abdul-Hakeem Shaibu <hkmshb@gmail.com>"]
keywords = ["hazel", "toolkit"]
license = "BSD-2-Clause"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.6"
"zope.interface" = {version = "^4.5", optional = true}


[tool.poetry.extras]
services = ["wired", "zope.interface"]


[tool.poetry.dev-dependencies]
pytest = "^3.5"
pytest-cov = "^2.5"
pre-commit = "^1.11"


[tool.poetry.scripts]


[tool.black]
line-length = 79
check = true
diff = true
skip-string-normalization = true


[tool.tox]
legacy_tox_ini = """
[tox]
skipsdist = true
envlist = lint, py36

[testenv]
passenv = HOME USER
whitelist_externals =
    bash
    poetry
skip_install = true
commands =
    bash ./prepare-toxenv.sh {envbindir} {toxinidir}
    poetry run py.test --cov-report term-missing --cov=hazel tests

[testenv:lint]
whitelist_externals =
    bash
    poetry
skip_install = true
deps =
    pylint
commands =
    bash ./prepare-toxenv.sh {envbindir} {toxinidir}
    pylint --rcfile={toxinidir}/.pylintrc {toxinidir}/src/hazel {toxinidir}/tests
"""
